SuggestionButton as own component. Forwardref from Button
